+++ src/constants.rs
@@ -1,11 +1,13 @@
+use crate::types::Currency;
+
 /// Value for one weapon.
-pub const ONE_WEAPON: i32 = 1;
+pub const ONE_WEAPON: Currency = 1;
 /// Value for one scrap metal.
-pub const ONE_SCRAP: i32 = ONE_WEAPON * 2;
+pub const ONE_SCRAP: Currency = ONE_WEAPON * 2;
 /// Value for one reclaimed metal.
-pub const ONE_REC: i32 = ONE_SCRAP * 3;
+pub const ONE_REC: Currency = ONE_SCRAP * 3;
 /// Value for one refined metal.
-pub const ONE_REF: i32 = ONE_REC * 3;
+pub const ONE_REF: Currency = ONE_REC * 3;
 
 /// Symbol for one key.
 pub const KEY_SYMBOL: &str = "key";