+++ src/error.rs
@@ -1,50 +1,25 @@
-use std::fmt;
 use std::num::{ParseFloatError, ParseIntError};
 
-#[derive(Debug)]
+/// Error converting listing currencies to currencies.
+#[derive(Debug, thiserror::Error)]
+#[error("Currencies contains fractional value: {fract}")]
 pub struct TryFromListingCurrenciesError {
+    /// Fractional key values are invalid.
     pub fract: f32,
 }
 
-impl std::error::Error for TryFromListingCurrenciesError {}
-
-impl fmt::Display for TryFromListingCurrenciesError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Currencies contains fractional value: {}", self.fract)
-    }
-}
-
-#[derive(Debug)]
+#[derive(Debug, thiserror::Error)]
 pub enum ParseError {
-    NoCurrencies,
+    /// String was invalid.
+    #[error("No currencies could be parsed from string")]
     Invalid,
+    /// A value expected to be number failed to parse. 
+    #[error(r#"Failed to parse "{}" as numeric"#, .0)]
     ParseNumeric(String),
-    ParseInt(ParseIntError),
-    ParseFloat(ParseFloatError),
-}
-
-impl From<ParseIntError> for ParseError {
-    fn from(error: ParseIntError) -> Self {
-        Self::ParseInt(error)
-    }
-}
-
-impl From<ParseFloatError> for ParseError {
-    fn from(error: ParseFloatError) -> Self {
-        Self::ParseFloat(error)
-    }
-}
-
-impl std::error::Error for ParseError {}
-
-impl fmt::Display for ParseError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            ParseError::NoCurrencies => write!(f, "No currencies could be parsed from string"),
-            ParseError::Invalid => write!(f, "Invalid currencies string"),
-            ParseError::ParseNumeric(e) => write!(f, "{}", e),
-            ParseError::ParseInt(e) => write!(f, "{}", e),
-            ParseError::ParseFloat(e) => write!(f, "{}", e),
-        }
-    }
+    /// A string failed to parse to an integer.
+    #[error("{}", .0)]
+    ParseInt(#[from] ParseIntError),
+    /// A string failed to parse to a float.
+    #[error("{}", .0)]
+    ParseFloat(#[from] ParseFloatError),
 }