+++ src/helpers.rs
@@ -1,4 +1,5 @@
 use crate::error::ParseError;
+use crate::types::Currency;
 use crate::constants::{KEYS_SYMBOL, KEY_SYMBOL, METAL_SYMBOL, ONE_REF};
 use crate::Rounding;
 use std::fmt;
@@ -7,54 +8,69 @@ use serde::{Deserialize, Deserializer};
 
 /// Converts currencies to a metal value using the given key price (represented as weapons). This
 /// method is saturating.
-pub fn to_metal(metal: i32, keys: i32, key_price: i32) -> i32 {
+pub fn to_metal(
+    metal: Currency,
+    keys: Currency,
+    key_price: Currency,
+) -> Currency {
     keys.saturating_mul(key_price).saturating_add(metal)
 }
 
 /// Converts currencies to a metal value using the given key price (represented as weapons).
-/// In cases where the result overflows or underflows beyond the limit for i32, `None` is 
+/// In cases where the result overflows or underflows beyond the limit for [`i64`], `None` is 
 /// returned.
-pub fn checked_to_metal(metal: i32, keys: i32, key_price: i32) -> Option<i32> {
+pub fn checked_to_metal(
+    metal: Currency,
+    keys: Currency,
+    key_price: Currency,
+) -> Option<Currency> {
     metal.checked_add(keys.checked_mul(key_price)?)
 }
 
-pub fn metal_deserializer<'de, D>(deserializer: D) -> Result<i32, D::Error>
+/// Deserializes float weapon values as weapons.
+pub fn metal_deserializer<'de, D>(deserializer: D) -> Result<Currency, D::Error>
 where
     D: Deserializer<'de>
 {
-    let float = f32::deserialize(deserializer)?;
-    let metal = (float * (ONE_REF as f32)).round() as i32;
+    // get the metal value as a float e.g. 2.55 ref
+    let metal_refined_float = f32::deserialize(deserializer)?;
+    // will fit it into the nearest weapon value
+    let metal = (metal_refined_float * (ONE_REF as f32)).round() as Currency;
     
     Ok(metal)
 }
 
+/// Serialzies and deserializes cents.
 pub mod cents {
     use serde::{Serializer, Deserialize, Deserializer};
+    use crate::types::Currency;
     use super::cents_to_dollars;
     
-    pub fn serialize<S>(value: &i32, serializer: S) -> Result<S::Ok, S::Error>
+    pub fn serialize<S>(value: &Currency, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer
     {
         serializer.serialize_f32(cents_to_dollars(*value))
     }
     
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<i32, D::Error>
+    pub fn deserialize<'de, D>(deserializer: D) -> Result<Currency, D::Error>
     where
         D: Deserializer<'de>
     {
         let usd = f32::deserialize(deserializer)?;
-        let cents = (usd * 100.0).round() as i32;
+        let cents = (usd * 100.0).round() as Currency;
         
         Ok(cents)
     }
 }
 
-pub fn cents_to_dollars(cents: i32) -> f32 {
+/// Converts cents to dollars.
+pub fn cents_to_dollars(cents: Currency) -> f32 {
     (cents as f32) / 100.0
 }
 
-pub fn pluralize<'a>(amount: i32, singular: &'a str, plural: &'a str) -> &'a str {
+/// Pluralizes a value using an integer as the test.
+pub fn pluralize<'a>(amount: Currency, singular: &'a str, plural: &'a str) -> &'a str {
     if amount == 1 {
         singular
     } else {
@@ -62,6 +78,7 @@ pub fn pluralize<'a>(amount: i32, singular: &'a str, plural: &'a str) -> &'a str
     }
 }
 
+/// Pluralizes a value using a float as the test.
 pub fn pluralize_float<'a>(amount: f32, singular: &'a str, plural: &'a str) -> &'a str {
     if amount == 1.0 {
         singular
@@ -70,12 +87,47 @@ pub fn pluralize_float<'a>(amount: f32, singular: &'a str, plural: &'a str) -> &
     }
 }
 
+/// Prints a float as either an integer if it contains no fractional values or with 2 decimal 
+/// places if it does.
 pub fn print_float(amount: f32) -> String {
-    if amount % 1.0 == 0.0 {
-        (amount.round() as i32).to_string()
+    if amount.fract() == 0.0 {
+        (amount.round() as Currency).to_string()
     } else {
         format!("{:.2}", amount)
     }
+}
+
+/// Adds thousands places to a number string e.g. "1000" becomes "1,000".
+pub fn thousands(string: String) -> String {
+    // from https://crates.io/crates/separator
+    let index = match string.find('.') {
+        Some(i) => i,
+        None => string.len()
+    };
+    let int_part = &string[..index];
+    let fract_part = &string[index..];
+    let mut output = String::new();
+    let magnitude = if let Some(stripped) = int_part.strip_prefix('-') {
+        output.push('-');
+        stripped.to_owned()
+    } else {
+        int_part.to_owned()
+    };
+    let mut place = magnitude.len();
+    let mut later_loop = false;
+    
+    for ch in magnitude.chars() {
+        if later_loop && place % 3 == 0 {
+            output.push(',');
+        }
+        
+        output.push(ch);
+        later_loop = true;
+        place -= 1;
+    };
+    
+    output.push_str(fract_part);
+    output
 }
 
 /// Converts a metal value into its float value.
@@ -85,7 +137,7 @@ pub fn print_float(amount: f32) -> String {
 /// ```
 /// assert_eq!(tf2_price::get_metal_float(6), 0.33);
 /// ```
-pub fn get_metal_float(value: i32) -> f32 {
+pub fn get_metal_float(value: Currency) -> f32 {
     f32::trunc((value as f32 / (ONE_REF as f32)) * 100.0) / 100.0
 }
 
@@ -96,11 +148,12 @@ pub fn get_metal_float(value: i32) -> f32 {
 /// ```
 /// assert_eq!(tf2_price::get_metal_from_float(0.33), 6);
 /// ```
-pub fn get_metal_from_float(value: f32) -> i32 {
-    (value * (ONE_REF as f32)).round() as i32
+pub fn get_metal_from_float(value: f32) -> Currency {
+    (value * (ONE_REF as f32)).round() as Currency
 }
 
-pub fn parse_from_string<T>(string: &str) -> Result<(T, i32), ParseError>
+/// Parses currencies from a string.
+pub fn parse_from_string<T>(string: &str) -> Result<(T, Currency), ParseError>
 where
     T: Default + FromStr + PartialEq,
     <T as FromStr>::Err: fmt::Display,
@@ -145,13 +198,14 @@ where
     }
     
     if keys == T::default() && metal == 0 {
-        return Err(ParseError::NoCurrencies);
+        return Err(ParseError::Invalid);
     }
     
     Ok((keys, metal))
 }
 
-pub fn round_metal(metal: i32, rounding: &Rounding) -> i32 {
+/// Rounds a metal value.
+pub fn round_metal(metal: Currency, rounding: &Rounding) -> Currency {
     if metal == 0 {
         return metal;
     }