+++ src/types.rs
@@ -1,0 +1,2 @@
+/// The integer type used for currencies.
+pub type Currency = i64;